Converting Jackson JsonParseException exception
In the case of a completely malformed JSON payload being submitted a JsonParseException was being thrown.  This change catches that exception and converts it into a ValidateJsonSchemaException exception which is more inline with the intent of how this library is intended to work.

This also fixes some of the deprecation warnings for methods that were using constructors to siwtch them to use builders.